Make Ohai Jruby compatible.
Use Process Builder instead of open4 to execute the plugin commands when in JRuby environment. Tweaked a few tests to make sure they do the right things in both MRI and JRuby.

Fix the gem spec so that we get a ohai-jruby gem or a ohai-gem based on how we run it. Also removed a useless puts