[OHAI-275] force encoding of command output to default_external
When reading a fixed number of bytes from an IO object, ruby 1.9 will
mark the strings as ASCII-8BIT (i.e., binary). The content, however,
should be encoded as the default_external encoding, usually UTF-8. This
causes great suffering when ruby tries to convert the mislabeled content
to UTF-8. So we force-set the encoding on the output strings to the
default_external encoding