use the ruby version that was used to run ohai
this fixes (spec) issues when multiple ruby versions are installed in
the system and the user can switch between them.

in gentoo specs are automatically run for all installed ruby versions,
but the global /usr/bin/ruby symlink points to the version the user has
selected.

if the user has selected ruby18 as default the ruby plugin will detect
ruby18 even if ohai is called with ruby19 explicitly.