[#OHAI-441] Restore EC2 '2008-02-01' userdata/metadata API support as a fallback
Older EC2 instances still only respond to the old API's. The current code takes
no notice of the response and tries to parse the 'Not Found' page. This reuslts
in a never ending recursive loop.

This fix adds checks for response codes, and falls back to the old '2008-02-01'
API if a request on the new API fails. It's a stateless setup, so it will always
try the new API first, then the old.