[#OHAI-434] Detect and use the latest (recognized-working) EC2 metadata API
Supported API's are determined at instance launch and are not extended
over the life of the instance. As such the current code - which makes an
assumption of a specific version - will fail depending on the age of the
instance.

This new code probes the instance metadata endpoint for available
versions, determines the most advanced version known to work and
executes the metadata retrieval using that version.

If no compatible version is found, the plugin returns an empty hash.