OHAI-445: Fallback to default HostKey path
This fixes a bug in the ssh_host_key plugin that prevented the plugin
from reading ssh_host_*_key files when a sshd_config exists.

When a sshd_config exists but the HostKey directive is not present, i.e.
using the default value, the plugin should populate the host_*_public
attributes using the default HostKey locations.