implement cpu count in solaris
the block of comments was not geting executed.

this uses kstat instead of psrinfo since google suggests that has
better backwards support to solaris 9/10 like we need.  we could
use psrinfo instead if it looks sane, but it looks like we'd need
to implemennt kstat as a fallback anyway...

i feel dirty using popen4, but everything else does as well..