Remove `force` option to run plugins
`Ohai::System#refresh_plugins` should be used instead.

This option was incompatible with the implementation of dependency graph
traversal in `Ohai::Runner`, and using it could lead to infinite loops
when running a plugin with a simple dependency structure.

The force option remains in the `require_plugin()` method to preserve
version 6 API compatibility, but it is ignored.