OHAI-550 java -version wastes memory
On newer versions of java the java heap can start as the smaller of 1/4 of memory or 1gb rather than 64mb like in the past. Even when just running -version. Protect against mmaping 1gb of memory for no reason by specifying a 16mb max heap. This avoids causing pain for boxes where memory is mostly used.