do not warn for missing gateway IPs
When using a /32 address with a point-to-point gateway, you can either
use iproute2's peer option in which case ohai detects the ipaddress
correctly. Alternatively one can configure the /32 address on the
network and setup a static route to the gateway:

$ ip a
3: lxc0: <BROADCAST,MULTICAST,UP,LOWER_UP> mtu 1500 qdisc noqueue state UP
    link/ether 00:1e:8c:f4:f0:b1 brd ff:ff:ff:ff:ff:ff
    inet 1.2.3.4/32 scope global lxc0

$ ip r
default via 5.9.41.225 dev lxc0
5.9.41.225 dev lxc0  scope link

Using this configuration ohai does the right thing but logs a warning.
This patch simply converts the log line to debug to hide this.