OHAI-541: return a response even when ec2 metadata is 404
During EC2 metadata extraction, a given node in the metadata tree
may result in a '404 Not Found' message, causing the entire EC2 plugin
to fail, and returning an empty object.

Since it's better to have most of the metadata returned, this allows
the 404 to continue, and others should continue to fail.