[ohai 6] Linux::filesystems should prefer lsblk to blkid when available
blkid has several bugs, and will miss a variety of things.

For starters, it'll never report on a device that has partitions even if the
parent is the thing with a filesystem on it. It has weird bugs like:

http://marc.info/?l=linux-ext4&m=119636478426226&w=2

lsblk, on the other hand, probes everything in /sys/block and reports
consistently.