Workaround for hyperic/sigar#48, throw on 0 pid
When proc_state is called for the second and all future times on a
single sigar instance, it fails to throw an exception, resulting in
invalid memory access during ohai reporting of process names. This
workaround causes an invalid pid (0) to be caught and prevent the
invalid access.  Closes #423 and will be unnecessary and should be
reverted when hyperic/sigar#48 is resolved.