Fix and extend CPU detection on FreeBSD
The comment block and the tests had an incorrect sample of what
dmesg.boot looked like.  I got 67 different people to send me samples of
what the output looks like, with about 100 samples total.  The Origin
line is always indented and there are never spaces around the equals
symbol.  I updated the regex to properly detect the existing fields and
also added family/model, which we gather on Linux.