Do not set macaddress if it is already set
Previously we almost always set macaddress to the value from the interface we
took ipaddress from, unless we didn't get any ipaddress, at which point we
would accept the macaddress from the interface we took ip6address from.

We're seeing ipv6 only hosts now, we need to be able to take macaddress from
an ip6address interface if that's the most logical choice so we can ignore
loopback or random/meaningless virtual interfaces that don't go anywhere.