Avoid creating the mash until after we attempt to shell out
Also be smarter about when we save the data we found.  Some of these plugins collect 1/2 dozen attributes, but only use them if version attribute was found. Seems to be a lot of cargo cutting since this is almost all the same code.

There's no point creating the mash when most of the time the lang isn't on the host. Also add debugging for what was just successfully run.