Detect paravirt amazon instances without hint files
Official AMIs ship with the ec2metadata binary which we can look for. With this change we should be detecting the majority of scenarios without a hint file.

has_ec2_mac has been renamed has_xen_mac since we're actually looking for a xen mac.  Due to that the attempt to connect to the metadata only needs to happen when we only find the mac address since that can be either AWS or Xen.  If we find dmi data, a hint, or the metadata binary we can safely assume the metadata endpoint will respond.

 I also consolidated the "no hints or signs of AWS" scenario into a single spec since it seems to make more sense that way.