solaris11_network: Handle solaris 11 zone interfaces
PHYSRUNNING is now a valid flag for an interface.

The Solaris 11 zones "route -n get default” command output returns the
interface name, net0, but not the specific instance net0:2. The output
changed from Solaris 10 to Solaris 11. Use the verbose output
"route -v -n get default" to find the interface index and use the
name and index to find the full interface name from the ifconfig -a
output.