Use escape_glob_dir instead of escape_glob
For ruby >= 2.2, Dir.glob behaves differently than ruby 2.0
and breaks in more places when backslashes are used when globbing.
This means that Windows is sad and will not load any plugins with
Ruby >= 2.2 in a lot of cases. This is the safest thing to do
because it will use forward slashes.