Comment cleanup
virtualization[:mechanism] would have been great, but we didn't setup the data structure right for this to work
hypervisor flag is present anywhere the guest has a hypervisor.  It's not just a Xen thing
If we find KVM CPUs we're on an old paravirt box and BTMs request for info from 2009 is probably not going to get answered. Sorry BTM