VirtualizationInfo plugin cleanup
This is an early plugin that hasn't received any work since it's creation.  I seriously doubt anyone is using it, but this provides some minor cleanup

1) remove the hpricot dependency and all the commented out hpricot code. This plugin would fail unless it was installed, but it's not actually used
2) Stop storing XML data on the node. This is a horrible horrible thing to do. We're storing the exact same data in JSON. Why 2 formats?
3) Change the debug message to make it clean what plugin fails if the gems aren't there
4) Add a comment telling anyone reading this what gem to use.  The libvirt gem is deprecated and fails to load.

It's a bummer we need this gem to use this because the data is actually pretty useful once it's loaded, but I doubt anyone installs the gem.