Don't test hint logic in every cloud plugin
We test this once in the hint spec.  We don't need to expect the windows and linux file paths.  Additionally if you're on a host that actually has a hint it breaks these tests.  The hint spec does it all in a temp dir which is the right way to test it.

Also while I'm in these specs update the format, fix some poor wording, and use shared_examples in a few places to dry things up