Exclude vendor data and don't depend on the network plugin
The network plugin is leftover from the old method where we looked up the interfaces based on the IP in the hint.  We don't do that anymore so this isn't necessary.

Vendor data can include sensitive bootstrapping information that we shouldn't save to the node.  It also doesn't include anything of value so either way the data shouldn't be on the node