Rework the specs
These specs weren’t covering the actual meat of the plugin. They mocked out the methods, when the methods are where the majority of logic exists. Instead of mocking those methods mock out the files on disk. This also serves to document the structure of the files we’re parsing in the plugin so that’s another win.

Signed-off-by: Tim Smith <tsmith@chef.io>