Move the “Virtualization” plugin into libvirt namespace
This plugin isn’t actually a virtualization plugin. It’s actually a libvirt plugin. Move the data out of the Virtualization mash and stick it into its own location. Also rename the plugin LibVirt. I need to get a system setup with libvirt to test this and probably tweak it further / add tests.

Signed-off-by: Tim Smith <tsmith@chef.io>