Make filesystem more resilient
Today lsblk and blkid each have their various shortcomings. Various hacks
and workarounds have been added and changed as we migrated to using lsblk
over blkid, but we can do better.

We run over both and combine the data.

This was caught when we noticed that somne of our AOE-type filesystems were
not recognized by lsblk when not mounted, but blkid did notice them. This meant
we thought the devices were not formatted and wanted to format them. That's bad.
Mega bad.

Signed-off-by: Phil Dibowitz <phil@ipom.com>