Critical Plugins
This adds a new Ohai configuration `critical_plugins` which will cause
Ohai to exit (or `fail` if inside of Chef) if the plugins in the list do
not run successfully.

This feature only supports Ohai 7-style named plugins, not the ancient
Ohai 6-style stuff.

Tests to follow, but:

```
$ cat /home/phild/ohai.rb
ohai.critical_plugins << :Filesystem
$ ./bin/ohai -c /home/phild/ohai.rb
[2017-10-03T13:43:14-07:00] INFO: The plugin path /etc/chef/ohai/plugins does not exist. Skipping...
[2017-10-03T13:43:14-07:00] ERROR: The following plugins marked as critical failed: [:Filesystem]
```

And when I run Chef:
```
[2017-10-03T13:44:11-07:00] FATAL: RuntimeError: The following plugins marked as critical failed: [:Filesystem2]. Failing Chef run.
```

This fixes #879.

Signed-off-by: Phil Dibowitz <phil@ipom.com>