Remove unused platform simulation specs
We never actually completed this work and started using these. It's just dead code in the repo. We do need integration testing on Ohai, but I'm not sure if this gets us much since we're doing the same sort of hack on shell out in our existing "unit" (yes in quotes) tests. What we could really use is a true integration test where we run ohai on various platforms and expect back certain output.

Signed-off-by: Tim Smith <tsmith@chef.io>