[linux/network] Tunnel information (#1104)
This gathers extra information on tunnel devices. Unfortunately iproute2 is a
little inconsistent as to how it prints these out, some are booleans some are
key values, but there's no way to tell them apart other than to know, so I parse
the ones I know about (from reading the source), and ignore anything else.

Note that I didn't add any ifconfig support to the tests. This is for a few
reasons:

1. the output in the tests from ifconfig is ancient, ifconfig on linux doesn't
look like that anymore, it looks like this:

```
eth0: flags=4163<UP,BROADCAST,RUNNING,MULTICAST>  mtu 1500
        inet 10.1.1.2  netmask 255.255.255.0  broadcast 10.1.1.255
        inet6 2601:645:c001:56bb::2  prefixlen 64  scopeid 0x0<global>
        inet6 fe80::52e5:49ff:fe38:d761  prefixlen 64  scopeid 0x20<link>
        ether 50:e5:49:38:d7:61  txqueuelen 1000  (Ethernet)
        RX packets 203501353  bytes 98738916585 (91.9 GiB)
        RX errors 0  dropped 559  overruns 0  frame 0
        TX packets 1035989265  bytes 1458478887585 (1.3 TiB)
        TX errors 0  dropped 0 overruns 0  carrier 0  collisions 0
```

And I couldn't find a machine old enough to generate the type of output in the
tests *and* have modern ip6tunnel drivers on it. So don't ask. :)

2. Only iproute2 shows the extra tunnel info anyway, so it's not really useful
to do have the interface in the ifconfig output.

Note I will backport this to Ohai 8 after this is merged.

Signed-off-by: Phil Dibowitz <phil@ipom.com>