Load collect_data() even if we've already seen it
If we happen to load a plugin with the same name twice, we refuse to run #collect_data but we still
load the rest of the plugin, which means you get #collect_data from the first plugin and rest of the
plugin from the last plugin.

This makes it so you get #collect_data from the last plugin as well.

I don't know how to use the logger from the class method so I fell back to `Ohai::Log`. Happy to fix that
if someone knows the magic.

Fixes #1212

Signed-off-by: Bryan McLellan <btm@chef.io>