Avoid gathering all data with sysctl which seems to hang
While it works fine for me @bdausses was seeing chef-client hang
forever. It turns out running sysctl -a on his host was causing the
system to hang. Besides this crash it actually takes about 2x longer to
gather the additional data. This shaves a whole 0.033255 seconds from
the total runtime.

Signed-off-by: Tim Smith <tsmith@chef.io>