Require Ruby 2.5 or later
As we do every year Ohai 15 will drop support for the N-2 support of Ruby. In the ~5 weeks Ruby 2.6 comes out so by the time we release this new version of Ohai N-2 will be Ruby 2.4.

Also while we're at it get rid of the begins in rescue blocks. Ruby 2.5 doesn't require this and it's on the path to ruby throwing errors eventually. It's a Ruby < 2.4-ism.

Signed-off-by: Tim Smith <tsmith@chef.io>