Unify virtualization detection on a single helper
We've always detected virtualzation on *nix using dmidecode data. A few
years ago I unified this with the dmi_decode helper. It parsed out the
text of the dmidecode command. This greatly improved detection on BSD
and Solaris, but there was more that could be done. With this change
we're now passing in the manufacturer, product, and version data from
the dmi data. On the *nix side we'll use the dmi plugins data which has
already been split out nicely and doesn't need to be fetched or parsed
again. On the windows side we can actually get this data via WMI pretty
easily. The field names are different, but under the hood it's just DMI
data. This way if we accept a PR from a windows, solaris, bsd, or linux
user we get improves on all the other platforms.

Note: This also removes PCI based Parallels detection on Linux since we're doing it in a more efficient way with DMI data now on Linux.

Signed-off-by: Tim Smith <tsmith@chef.io>