Fixes the file_in_dir and re-use executable?
To replace Dir[] I built file_in_dir but I don't at this
point want to provide the same [] interface. So I changed
it to a method invocation.

Changed the the file_open back to return the StringIO object
I think that there was some other issue that was leading me
away from it when inside a block. Also I forgot that when you
examine each line of the StringIO you need to reset the cursor
that is keeping track of the position.

Decided to reuse the file_executable? in the which helper
which improves both.

Made notes about lacking windows support

Signed-off-by: Franklin Webber <franklin.webber@gmail.com>