Moves the WhichHelper to added at runtime
When the backend is windows load the windows which helper

I don't like this approach because it violates some of the ability to
see what is happening in the plugin. I think there may be a better
approach where the which always checks for the right which or loads the
right which the first time. But ultimately a plugin probably doesn't do
a lot of whiching.

Signed-off-by: Franklin Webber <franklin.webber@gmail.com>