Merge the new vbox plugin into the existing virtualbox plugin
It makes sense to handle both guest and host logic in the same plugin since you can't be both. This gives a more predictable namespace and makes the code slightly smaller. I also removed the ostypes field since that doesn't change by system and is an enormous amount of data to collect with pretty limited value from a cookbook standpoint.

Signed-off-by: Tim Smith <tsmith@chef.io>