Gather CPU info from lscpu and create cpu/lscpu attribute namespace
This adds a cpu/lscpu node data name space containing information from the lscpu
command (if it exists). This augments and standardizes the information you can
gather from /proc/cpuinfo which can be an issue on non-x86 architectures such as
ppc64le and aarch64. Namely, when trying to determine whether or not a node is a
guest or not.

In addition, this properly fixes cpu counts on s390x nodes and uses lscpu data
for determining cpu/total, cpu/real and cpu/cores if lscpu exists.

Signed-off-by: Lance Albertson <lance@osuosl.org>