Fixes for detecting KVM guests on certain platforms and architectures
On some architectures (i.e. ppc64le), dmidecode does not exist nor does the
proper information in /proc/cpuinfo exist to detect KVM guests. To workaround
that, we can use lscpu data to easily determine this as a fallback.

In addition on CentOS 6, /sys/devices/virtual/misc/kvm does not exist so this
works around that problem as well.

Signed-off-by: Lance Albertson <lance@osuosl.org>