Windows support for Passwd plugin
This adds Windows support for the passwd plugin.

It's not fast, but I did work to improve the performance dropping it
from 20 seconds to 8 seconds on a VM with only a few users and groups.
The `get-localgroupmember` is the slow part (you have to run it one
group at a time), so the more groups you have, the slower this will go.

Note that unlike the Linux/Mac variety which will pick up nonlocal users
and groups this one does not (yet). It isn't hard to add, there are the
`get-aduser` and `get-adgroup` commands (if you have joined a domain),
but this is a good start.

And yes, I'll add tests when I have a moment.

Signed-off-by: Phil Dibowitz <phil@ipom.com>