[linux/platform] fix platform_version on Debian testing and unstable
When this plugin is updated to use `/etc/os-release`, it introduces a regression
on Debian testing and unstable, where the `VERSION_ID` field is not set.

In the legacy code path, `platform_version` is set on Debian by reading
`/etc/debian_version`; this is still needed on Debian testing and unstable.

Add a special case for Debian, falling back to reading this file rather than
to using `uname -r`, and update the spec to cover this.

Signed-off-by: Michel Alexandre Salim <michel@fb.com>