Add support for Podman and fix docker guest detection on Github Actions
This provides proper support for detecting Podman hosts and guests which is now
the default container system on Fedora and RHEL 8+.

In addition, this works around a problem when running containers on Github
Actions. They apparently configure the docker on their runs in a non-standard
way which renames the cgroup mounts to `actions_job`.

It also shows `/system.slice/containerd.service` as a cgroup mount point and
should be a safe assumption this is running inside a docker container. This
should be checked last in the logic order since it would be triggered in the
first check if it were added there.

Signed-off-by: Lance Albertson <lance@osuosl.org>