Merge data from lscpu into cpu attributes on Linux
If lscpu exists and returns data, the cpu data will be filled using that data
(with a few exceptions). If lscpu does not return data, it will fall back to the
old method of only reading data from /proc/cpuinfo.

Some data on a few architectures is only available from /proc/cpuinfo, so add it
when needed when using lscpu.

In addition, this changes the unit tests to pull data from files instead of
adding it directly to the unit test file.

Signed-off-by: Lance Albertson <lance@osuosl.org>