Ohai: Change C language plugin for glibc version detection
It's not safe to directly execute "libc.so.6" to determine glibc version.
This can lead to segfaults when other shared libraries are preloaded using
LD_PRELOAD or /etc/ld.so.preload due to circular relocation dependencies.

Replace this version detection with "ldd --version" as "ldd" is a core part
of glibc and reports the same version.

Signed-off-by: Bhavesh Davda <bhavesh.davda@oracle.com>