Gather CPU info from lscpu and merge data into cpu attribute namespace (#1454)
* Gather CPU info from lscpu and create cpu/lscpu attribute namespace

This adds a cpu/lscpu node data name space containing information from the lscpu
command (if it exists). This augments and standardizes the information you can
gather from /proc/cpuinfo which can be an issue on non-x86 architectures such as
ppc64le and aarch64. Namely, when trying to determine whether or not a node is a
guest or not.

In addition, this properly fixes cpu counts on s390x nodes and uses lscpu data
for determining cpu/total, cpu/real and cpu/cores if lscpu exists.

* Fixes for detecting KVM guests on certain platforms and architectures

On some architectures (i.e. ppc64le), dmidecode does not exist nor does the
proper information in /proc/cpuinfo exist to detect KVM guests. To workaround
that, we can use lscpu data to easily determine this as a fallback.

In addition on CentOS 6, /sys/devices/virtual/misc/kvm does not exist so this
works around that problem as well.

Signed-off-by: Lance Albertson <lance@osuosl.org>