Add new rpm plugin
Add a basic plugin to expose configuration information from the RPM
package manager. I specifically wrote this to expose the `_db_backend`
macro, but figured it'd be useful to make it cover a bit more while at
it.

This plugin is enabled on all non-Windows platforms, because a port of
RPM exists (and is sometimes used) on pretty much all platforms, even
where it's not usually installed by default.

Signed-off-by: Davide Cavalca <dcavalca@fb.com>