networking/linux: map src only default routes accordingly
in certain configurations a default route can be set via src only / not
dev, but should still be accounted for in terms of calculating the
'favored' default route / what is chosen as ipaddress for a given node.

This change augments the previously added changes to account for nexthop
multipath routing; but only in scenarios src address is set but only in
scenarios src address is set.

Signed-off-by: Matt Phillips <mphillips81@bloomberg.net>